pull to 1.1

+++ NoseBot/Modules/TwitchMonitor.cs
@@ -48,11 +48,23 @@ namespace NoseBot.Modules
             string guildid = Context.Guild.Id.ToString();
             string guildpath = FileDirUtil.GetGuildDir(guildid);
             string startfile = Path.Combine(guildpath, FileDirUtil.JSONSTART);
+            string stopfile = Path.Combine(guildpath, FileDirUtil.JSONSTOP);
+
+            //if already running
             if (File.Exists(startfile))
             {
                 await ReplyAsync("Bot already running, if it appears to be broken try to restart");
                 return;
             }
+            //if a stop command already issued but not completed, don't try to re-start
+            if (File.Exists(stopfile))
+            {
+                await ReplyAsync("Awaiting termination of the existing session before restarting, please try again shortly!");
+                return;
+            }
+
+            //if not already running + no stop command in progress, start the bot
+
             FileStream fs = File.Create(startfile);
             fs.Flush();
             fs.Close();
@@ -75,12 +87,21 @@ namespace NoseBot.Modules
             string guildid = Context.Guild.Id.ToString();
             string guildpath = FileDirUtil.GetGuildDir(guildid);
             string startfile = Path.Combine(guildpath, FileDirUtil.JSONSTART);
+            string stopfile = Path.Combine(guildpath, FileDirUtil.JSONSTOP);
+            FileStream fs = File.Create(stopfile);
+            fs.Flush();
+            fs.Close();
+            if (File.Exists(stopfile))
+            {
+                await ReplyAsync("Waiting for previous session to end, please try again shortly");
+            }
             if (File.Exists(startfile))
             {
                 try
                 {
                     File.Delete(startfile);
                     await ReplyAsync("Monitor has been stopped! Please wait up to 1 minute before attempting a restart..");
+                    Console.WriteLine("stopped");
                 }
                 catch (Exception e)
                 {
@@ -247,6 +268,21 @@ namespace NoseBot.Modules
                 }
                 
             }
+
+            //once the monitor stops, remove the stop file
+            string guildid = Context.Guild.Id.ToString();
+            string guildpath = FileDirUtil.GetGuildDir(guildid);
+            try
+            {
+                File.Delete(Path.Combine(guildpath, FileDirUtil.JSONSTOP));
+                Console.WriteLine("removed stop file");
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine("could not remove stop file: " + e);
+
+            }
+            
             await ReplyAsync("**Monitor stopped**");
         }
 
@@ -262,10 +298,10 @@ namespace NoseBot.Modules
             Dictionary<string, string> live = JSONUtil.GetJsonToDic<string, string>(guildlive);
             List<string> liveusers = new List<string>();
 
-            string s = string.Join(",", users.Select(x => x.Value));
+            string s = string.Join("&channel=", users.Select(x => x.Value));
 
             string response = SendRequest(BaseString + "streams/?channel=", s, "Requesting live channels for " + s);
-
+            Console.WriteLine("channelresponse" + response);
             JSONUtil.WriteJsonToFile(null, guildstream, response);
 
             Streams streams = JsonConvert.DeserializeObject<Streams>(response);
@@ -291,10 +327,12 @@ namespace NoseBot.Modules
                 }
             }
 
+            Console.WriteLine("check1");
             //check the current live against previous live
             //if set as live when not currently live, set to offline
             Dictionary<string,string> newDictionary = live.ToDictionary(entry => entry.Key,
                                                entry => entry.Value);
+            Console.WriteLine("check2");
             foreach (KeyValuePair<string, string> entry in live)
             {
                 if (entry.Value == "live")
@@ -305,10 +343,12 @@ namespace NoseBot.Modules
                     }
                 }
             }
+            Console.WriteLine("check3");
             live = newDictionary.ToDictionary(entry => entry.Key,
                                                entry => entry.Value);
+            Console.WriteLine("check4");
             JSONUtil.WriteJsonToFile(live, guildlive);
-            Console.WriteLine("5");
+            Console.WriteLine("Completed check\n");
 
         }
 
@@ -318,7 +358,7 @@ namespace NoseBot.Modules
             var client = new RestClient(url + suffix);
             var request = new RestRequest(Method.GET);
 
-            request.AddHeader("Client-ID", "uo6dggojyb8d6soh92zknwmi5ej1q2");
+            request.AddHeader("Client-ID", "vm65hok0u8pse4s6h8s7f5ecvl20yt");
             request.AddHeader("Accept", "application/vnd.twitchtv.v5+json");
             request.AddParameter("application/vnd.twitchtv.v5+json", "bodykey=bodyval", ParameterType.RequestBody);
